DO NOT SUBMIT
This call will call the /formlogin URL with post and that works, however
the angular side can't tell if it's logged in since I can't figure out how
to get he scope variable I added to work on the page.

Also the js was missing under the dist directory, so that wasn't working
right, added that.  the css is still missing under dist.

For testing I copied everything under dist to the spring boot
webapp/ directory in our tree and ran it all through spring's
embedded tomcat. Worked great.